rename EditPage as TourPage
