+++ resources/camino-creator/pages/TourPage/InitialLocation.vue
