+++ resources/camino-creator/pages/TourPage/SelectTourStyle.vue
