+++ resources/camino-creator/pages/TourPage/ShareTour.vue
