+++ resources/camino-creator/pages/TourPage/TourPage.vue
