+++ resources/camino-creator/pages/TourPage/TourStopCard.vue
